remove inventory state from url
